Decompose widgets folder and remove iconColor field
