+++ packages/talker_flutter/lib/src/src.dart
@@ -5,4 +5,4 @@ export 'ui/talker_listener.dart';
 export 'ui/talker_screen.dart';
 export 'ui/talker_wrapper/talker_wrapper.dart';
 export 'ui/theme/theme.dart';
-export 'ui/widgets/cards/talker_data_card.dart';
+export 'ui/widgets/data_card.dart';
