+++ packages/talker_flutter/lib/src/ui/theme/talker_screen_theme.dart
@@ -5,7 +5,6 @@ class TalkerScreenTheme {
   const TalkerScreenTheme({
     this.backgroudColor = const Color(0xFF212121),
     this.textColor = Colors.white,
-    this.iconsColor = Colors.white,
   });
 
   /// Background screen color
@@ -13,7 +12,4 @@ class TalkerScreenTheme {
 
   /// Color of text on screen
   final Color textColor;
-
-  /// Colors of all Talker widgets icons
-  final Color iconsColor;
 }
