+++ packages/talker_flutter/lib/src/ui/ui.dart
@@ -1,2 +1,1 @@
 export 'theme/theme.dart';
-export 'widgets/widgets.dart';
