--- packages/talker_flutter/lib/src/ui/widgets/cards/cards.dart
@@ -1,1 +1,0 @@
-export 'talker_data_card.dart';
