+++ packages/talker_flutter/lib/src/ui/widgets/data_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:talker_flutter/src/ui/theme/default_theme.dart';
-import 'package:talker_flutter/src/ui/widgets/cards/base_card.dart';
+import 'package:talker_flutter/src/ui/widgets/base_card.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
 class TalkerDataCard extends StatelessWidget {
