--- packages/talker_flutter/lib/src/ui/widgets/widgets.dart
@@ -1,1 +1,0 @@
-export 'cards/cards.dart';
